Workaround for sample joke not being generated
