+++ .github/workflows/build.yml
@@ -94,6 +94,7 @@ docker push $INDEX_IMAGE
 
 # Apply RBAC
 kubectl apply -f manifests/jokereconciler.csv.rbac.yml -n $K8S_NAMESPACE
+kubectl apply -f manifests/jokerequestreconciler.csv.rbac.yml -n $K8S_NAMESPACE
 
 cat <<EOF | kubectl apply -f -
 apiVersion: operators.coreos.com/v1alpha1
@@ -121,15 +122,16 @@ spec:
   sourceNamespace: $K8S_NAMESPACE
 EOF
 
+retries=30
 until [[ $retries == 0 ]]; do
-  new_csv_phase=$(kubectl get csv -n "$K8S_NAMESPACE" jokereconciler -o jsonpath='{.status.phase}' 2>/dev/null || echo "Waiting for CSV to appear")
+  new_csv_phase=$(kubectl get csv -n "$K8S_NAMESPACE" jokerequestreconciler -o jsonpath='{.status.phase}' 2>/dev/null || echo "Waiting for CSV to appear")
   if [[ "$new_csv_phase" == "Succeeded" ]]; then
     break
-  elif [[ $retries == 30 ]]; then
-    echo "CSV \"jokereconciler\" failed to reach phase succeeded"
+  elif [[ $retries == 1 ]]; then
+    echo "CSV \"jokerequestreconciler\" failed to reach phase succeeded"
     exit 1
   else
-    echo "CSV \"jokereconciler\" phase: $new_csv_phase"
+    echo "CSV \"jokerequestreconciler\" phase: $new_csv_phase"
   fi
   sleep 10s
   retries=$((retries - 1))
@@ -139,10 +141,22 @@ cat <<EOF | kubectl create -f -
 apiVersion: "samples.javaoperatorsdk.io/v1alpha1"
 kind: JokeRequest
 metadata:
-  name: jr-any
+  name: jr-request
   namespace: $K8S_NAMESPACE
 spec:
   category: Any
 EOF
 
-# TODO: Check response from operator
+# Check response from operator
+retries=30
+until [[ $retries == 0 ]]; do
+  new_joke=$(kubectl get joke -n "$K8S_NAMESPACE" 2>/dev/null || echo "Waiting for Joke to appear")
+  if [[ "$new_joke" =~ .*"NAME".* ]]; then
+    break
+  elif [[ $retries == 1 ]]; then
+    echo "Failed to get jokes"
+    exit 1
+  fi
+  sleep 10s
+  retries=$((retries - 1))
+done
